requirements.txt
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
flask==3.0.0
flask-cors==4.0.0
openpyxl==3.1.2
xlsxwriter==3.1.9
pyyaml==6.0.1
scikit-learn==1.3.2
scipy==1.11.4
nltk==3.8.1
beautifulsoup4==4.12.2
jinja2==3.1.2
reportlab==4.0.8
python-dateutil==2.8.2
click==8.1.7
schedule==1.2.0
pytest==7.4.3
python-dotenv==1.0.0
gunicorn==21.2.0
redis==5.0.1